requirements.txt
numpy
torch
torchvision
tqdm
networkx
pyyaml
pillow
matplotlib
jupyter
ipython
nbconvert